feat: project done

+++ Controllers/PermissionController.cs
@@ -10,20 +10,29 @@ namespace FinalProject.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class PermissionsController(IPermissionsService permissionService) : ControllerBase
+    public class PermissionController(IPermissionService permissionService) : ControllerBase
     {
 
         [HttpGet]
-        [Authorize(Policy = "RequireAdminPermission")]
-        public async Task<ActionResult<List<Permission>>> GetAllPermissions()
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
+        public async Task<ActionResult<List<PermissionResponse>>> GetAllPermissions()
         {
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
             return Ok(await permissionService.GetAllPermissions());
         }
 
         [HttpPost]
-        [Authorize(Policy = "RequireAdminPermission")]
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
         public async Task<ActionResult<PermissionResponse>> CreatePermission(PermissionRequest request)
         {
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
+
             if (request == null) 
                 return BadRequest(request);
 
@@ -31,19 +40,29 @@ namespace FinalProject.Controllers
         }
 
         [HttpGet("{id}")]
-        [Authorize(Policy = "RequireAdminPermission")]
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
         public async Task<ActionResult<PermissionResponse>> GetPermissionById(int id)
         {
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
+
             var result = await permissionService.GetPermissionById(id);
             if (result == null) return NotFound($"Permission with id {id} not found");
             return Ok(result);
         }
 
         [HttpPut("{id}")]
-        [Authorize(Policy = "RequireAdminPermission")]
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
         public async Task<ActionResult<PermissionResponse>> UpdatePermission(int id, PermissionRequest request)
         {
-            if(await permissionService.GetPermissionById(id)==null)
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
+
+            if (await permissionService.GetPermissionById(id)==null)
             return BadRequest($"Permission with id {id} not found");
 
             var result = await permissionService.GetAllPermissions();
@@ -53,25 +72,42 @@ namespace FinalProject.Controllers
         }
 
         [HttpDelete("{id}")]
-        [Authorize(Policy = "RequireAdminPermission")]
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
         public async Task<ActionResult> DeletePermission(int id)
         {
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
+
+            var result = await permissionService.GetPermissionById(id);
+            if (result == null) return NotFound($"Permission with id {id} not found");
             await permissionService.DeletePermission(id);
             return Ok($"Permission with id {id} was deleted!");
         }
 
-        [HttpPost("~/api/AssignPermissionToUser")]
-        [Authorize(Policy = "RequireAdminPermission")]
+        [HttpPut("~/api/AssignPermissionToUser")]
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
         public async Task<ActionResult> AssignPermissionToUser(int userId,int permissionId)
         {
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
+
             await permissionService.AssignPermissionToUser(userId, permissionId);
             return Ok($"Permission with id {permissionId} has been assigned to User with Id {userId} successfully!");
         }
 
         [HttpPut("~/api/RemovePermissionForUser")]
-        [Authorize(Policy = "RequireAdminPermission")]
+        [Authorize]
+        //[Authorize(Policy = "RequireAdminPermission")]
         public async Task<ActionResult> RemovePermissionForUser(int userId, int permissionId)
         {
+            var permissions = User.Claims.Where(i => i.Type == "Permission").Select(i => i.Value).ToList();
+            if (!permissions.Contains("ADMIN"))
+                return Forbid();
+
             await permissionService.RemovePermissionForUser(userId, permissionId);
             return Ok($"Permission with id {permissionId} has been removed from User with Id {userId} successfully!");
         }
